requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
Django==5.0
django-appconf==1.0.6
django-compressor==4.4
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.3
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
lz4==4.3.3
moviepy==1.0.3
numpy==1.26.3
openpyxl==3.1.2
pandas==2.1.4
pillow==10.2.0
proglog==0.1.10
psutil==5.9.8
pydub==0.25.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3.post1
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.4
urllib3==2.2.0
